修改oauth接口报错
