+++ config/config.default.js
@@ -2,7 +2,7 @@
  * @Author: 杨仕明 shiming.y@qq.com
  * @Date: 2024-02-17 12:40:34
  * @LastEditors: 杨仕明 shiming.y@qq.com
- * @LastEditTime: 2024-03-22 17:23:30
+ * @LastEditTime: 2024-03-23 02:07:17
  * @FilePath: /Lulab_backend/config/config.default.js
  * @Description:
  *
